# frontend/api/openapi.json
{
  "components": {
    "schemas": {
      "ActualizarCargoRequestDto": {
        "properties": {
          "descripcion": {
            "type": "string"
          },
          "nombre": {
            "minLength": 1,
            "type": "string"
          },
          "rolIds": {
            "items": {
              "format": "int32",
              "type": "integer"
            },
            "type": "array"
          }
        },
        "required": [
          "nombre",
          "rolIds"
        ],
        "type": "object"
      },
      "ActualizarTrabajadorRequestDto": {
        "properties": {
          "apellidos": {
            "minLength": 1,
            "type": "string"
          },
          "cargoId": {
            "format": "int32",
            "type": "integer"
          },
          "celular": {
            "type": "string"
          },
          "correo": {
            "format": "email",
            "type": "string"
          },
          "nombres": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "apellidos",
          "cargoId",
          "nombres"
        ],
        "type": "object"
      },
...
      "BootstrapAdminRequestDto": {
        "properties": {
          "apellidos": {
            "minLength": 1,
            "type": "string"
          },
          "celular": {
            "minLength": 1,
            "type": "string"
          },
          "correo": {
            "format": "email",
            "minLength": 1,
            "type": "string"
          },
          "dni": {
            "minLength": 1,
            "type": "string"
          },
          "nombres": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          },
          "username": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "apellidos",
          "celular",
          "correo",
          "dni",
          "nombres",
          "password",
          "username"
        ],
        "type": "object"
      },
      "CambiarPasswordRequestDto": {
        "properties": {
          "passwordNueva": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "passwordNueva"
        ],
        "type": "object"
      },
...
      "CargoResponseDto": {
        "properties": {
          "descripcion": {
            "type": "string"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "nombre": {
            "type": "string"
          },
          "rolesNombres": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "CrearCargoRequestDto": {
        "properties": {
          "descripcion": {
            "type": "string"
          },
          "nombre": {
            "minLength": 1,
            "type": "string"
          },
          "rolIds": {
            "items": {
              "format": "int32",
              "type": "integer"
            },
            "type": "array"
          }
        },
        "required": [
          "nombre"
        ],
        "type": "object"
      },
...
      "PageResponseDtoTrabajadorResponseDto": {
        "properties": {
          "contenido": {
            "items": {
              "$ref": "#/components/schemas/TrabajadorResponseDto"
            },
            "type": "array"
          },
          "pagina": {
            "format": "int32",
            "type": "integer"
          },
          "primera": {
            "type": "boolean"
          },
          "tamano": {
            "format": "int32",
            "type": "integer"
          },
          "totalElementos": {
            "format": "int64",
            "type": "integer"
          },
          "totalPaginas": {
            "format": "int32",
            "type": "integer"
          },
          "ultima": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "Pageable": {
        "properties": {
          "page": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "sort": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "RegistrarTrabajadorRequestDto": {
        "properties": {
          "apellidos": {
            "minLength": 1,
            "type": "string"
          },
          "cargoId": {
            "format": "int32",
            "type": "integer"
          },
          "celular": {
            "minLength": 1,
            "type": "string"
          },
          "correo": {
            "format": "email",
            "type": "string"
          },
          "dni": {
            "minLength": 1,
            "type": "string"
          },
          "nombres": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          },
          "username": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "apellidos",
          "cargoId",
          "celular",
          "dni",
          "nombres",
          "password",
          "username"
        ],
        "type": "object"
      },
...
      "TrabajadorResponseDto": {
        "properties": {
          "activo": {
            "type": "boolean"
          },
          "apellidos": {
            "type": "string"
          },
          "cargoId": {
            "format": "int32",
            "type": "integer"
          },
          "cargoNombre": {
            "type": "string"
          },
          "celular": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "dni": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "nombres": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API de POS, inventario, comandas, cocina, despacho y caja del sistema de gestion gastronomica Chaquena.",
    "title": "Backend Logistica - Chaquena",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
    "/api/v1/auth/bootstrap": {
      "post": {
        "operationId": "bootstrap",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapAdminRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajadorResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Crear el primer administrador (solo con la tabla de trabajadores vacia)",
        "tags": [
          "Autenticacion"
        ]
      }
    },
...
    "/api/v1/cargos": {
      "get": {
        "operationId": "listarCargos",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CargoResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Cargos definidos con sus roles",
        "tags": [
          "Cargos"
        ]
      },
      "post": {
        "operationId": "crearCargo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearCargoRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Crear cargo con los roles que agrupa",
        "tags": [
          "Cargos"
        ]
      }
    },
    "/api/v1/cargos/{id}": {
      "get": {
        "operationId": "obtenerCargo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Obtener cargo por id",
        "tags": [
          "Cargos"
        ]
      },
      "put": {
        "description": "Quien ya inicio sesion conserva sus permisos hasta que vuelva a entrar: los roles viajan dentro del JWT.",
        "operationId": "actualizarCargo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarCargoRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Cambiar el nombre y el conjunto de roles del cargo",
        "tags": [
          "Cargos"
        ]
      }
    },
...
    "/api/v1/trabajadores": {
      "get": {
        "operationId": "listarTrabajadores",
        "parameters": [
          {
            "in": "query",
            "name": "cargoId",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageable",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoTrabajadorResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Listar trabajadores",
        "tags": [
          "Trabajadores"
        ]
      },
      "post": {
        "operationId": "registrarTrabajador",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrarTrabajadorRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajadorResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Registrar trabajador",
        "tags": [
          "Trabajadores"
        ]
      }
    },
...
    "/api/v1/trabajadores/{id}": {
      "get": {
        "operationId": "obtenerTrabajador",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajadorResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Obtener trabajador por id",
        "tags": [
          "Trabajadores"
        ]
      },
      "put": {
        "operationId": "actualizarTrabajador",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarTrabajadorRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajadorResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Actualizar datos y cargo del trabajador",
        "tags": [
          "Trabajadores"
        ]
      }
    },
    "/api/v1/trabajadores/{id}/activo": {
      "patch": {
        "operationId": "cambiarActivoTrabajador",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "activo",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajadorResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Dar de alta o de baja al trabajador",
        "tags": [
          "Trabajadores"
        ]
      }
    },
    "/api/v1/trabajadores/{id}/password": {
      "patch": {
        "description": "No pide la contrasena anterior: es la puerta que usa administracion cuando alguien pierde el acceso a mitad de turno.",
        "operationId": "restablecerPasswordTrabajador",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarPasswordRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrabajadorResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Restablecer la contrasena de un trabajador",
        "tags": [
          "Trabajadores"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}